+++ config/settings/base.py
@@ -114,6 +114,8 @@ LOCAL_APPS = [
     'roundabout.cruises',
     'roundabout.calibrations',
     'roundabout.configs_constants',
+    'roundabout.search',
+    'roundabout.exports',
 ]
 # https://docs.djangoproject.com/en/dev/ref/settings/#installed-apps
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
