+++ roundabout/calibrations/urls.py
@@ -10,5 +10,4 @@ urlpatterns = [
     path('ajax/value/edit/<int:pk>/', view=views.ValueSetValueUpdate.as_view(), name='valueset_value_update'),
     path('ajax/calname/add/<int:pk>/', view=views.PartCalNameAdd.as_view(), name='part_calname_add'),
     path('ajax/eventreview/delete/<int:pk>/<int:user_pk>/', view=views.event_review_approve, name='event_review_delete'),
-    path('export/<int:pk>/', view=views.ExportCalibrationEvent.as_view(), name='export_calibration'),
 ]
